requirements.txt
alembic==1.4.3
aniso8601==6.0.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
aws-shell==0.2.2
awscli==1.19.24
Babel==2.9.0
backcall==0.2.0
boto3==1.17.6
botocore==1.20.24
Choco==1.0.5
Click==7.0
colorama==0.4.3
configobj==5.0.6
decorator==4.4.2
distlib==0.3.1
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
iniconfig==1.1.1
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
pluggy==0.13.1
postgres==3.0.0
prompt-toolkit==1.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2019.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.4
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.2.2
wcwidth==0.2.5
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.3.3
